Updated MongoDB URI and obscured it with environmental variable, merged schema into one file called userLibraryModels
